+++ test/functional-tests/node-transformers/preparing-transformers/comments-transformer/fixtures/combined-words-2.js
@@ -6,5 +6,4 @@
 //abc
 var test = 1;
 //cde
-//javascript-obfuscator
 /** @preserved */