+++ test/functional-tests/node-transformers/preparing-transformers/obfuscating-guards/conditional-comment-obfuscating-guard/fixtures/comments-transformer-interaction.js
@@ -1,0 +1,9 @@
+(function () {
+    var foo = 1;
+
+    // javascript-obfuscator:disable
+    var bar = 2;
+
+    // javascript-obfuscator:enable
+    var baz = 3;
+})();